impr: refactor instruction decoding
Add an `Instruction` struct which represents an executable instruction,
as well as the address of its operand (if any).
Refactor instruction decoding into a separate function. This is
beneficial as we can now separately invoke the instruction decoding and
execution steps.

Add an `AddressingMode` enum to represent the possible addressing modes
on the NES.
Move addressing mode documentation to the `AddressingMode` enum.

Add `Clone` derive to `Opcode`.