feat: add parsing for cartridge headers
Add a "rom" module containing the definitions for the `ROM` and `Header`
structs which represent a NES cartridge.
Add functions for parsing byte strings into `ROM` or `Header` structs
using nom.
Header parsing should be compatible with the official iNES
specification. Parsing for the rest of the cartridge is not currently
implemented.
Add tests to parse an example header and the header of a test ROM file.

Update executable to load the requested ROM, parse the header, and
display the header debug info to the user.

Add "iNES" and "PlayChoice" to the list of strings ignored by clippy
when linting documentation.