wip: implement NROM mapper functionality
Move the `Mapper` code to a submodule.
Add an NROM mapper module.

Replace the `Mapper` trait with a `MemoryType` enum for representing
different types of memory (e.g. CHR/PRG, ROM/RAM, etc.), along with
`Fetch` and `Store` traits for addressing memory in a `MemoryType`
variant.
Rewrite the `Mapper` trait itself to also only accomodate `fetch` and
`store` instructions, with the proper handling of different memory types
being delegated to implementors.

Add a `mapper_from_rom` function which creates a `Mapper` based on the
mapper ID specified in the provided ROM.

Implement a mostly correct constructor for the NROM mapper.
Implement the re-defined `Mapper` trait for `NROM`.
Add tests for `NROM` mapper memory addressing.
Ensure `Default` instances of the `CPU` struct are initialized with an
empty `NROM` mapper.

Modify the `NESMemory` struct to store a `Mapper` impl instance and an
`APU` instance directly, for memory-mapped I/O addressing purposes.
Configure `NESMemory` to use the `NROM` mapper by default.
Implement memory store/fetch operations for the cartridge and APU memory
spaces.

Add an `EmptyMapper` mapper for the purpose of debugging open bus
behaviour.
Add a test for open bus behaviour.

Add a construct and various field getters/setters for `ROM`.

Add a `new_vec` function which creates a `Vec` of a specified size and
fills it with clones of a provided value; that is, this allows us to
make non-sparse `Vec`s.

Add a very simple, unfinished APU implementation, which currently only
allows you to read/write to the "Status" register of the APU (and even
then the operations are technically no-ops).